+++ examples/base_cc_lb/variables.tf
@@ -241,9 +241,8 @@ variable "lb_enabled" {
   default     = true
 }
 
-variable "encryption_at_host_setting" {
-  type        = string
+variable "encryption_at_host_enabled" {
+  type        = bool
   description = "User input for enabling or disabling host encryption"
-  default     = false
-
+  default     = true
 }